backport from trunk to 1.1.x release branch
if you only have a classic SE, there will still be a [se] section with the default_se option.  This will result in a GlueService stanza being written with garbage information in it. So, check for default_se = "UNKNOWN" or "UNAVAILABLE" and return if it does

default_se is set in the [Storage] section in config.ini and is required by configure-osg.py