GIP will now look in the config files for the proper user for dropping privileges
gip_info no longer attempts to drop privileges.  It would always fail since the import of osg_info_wrapper.py triggers the privilege drop in that module.

osg_info_wrapper.py checks the config object for the proper username, then checks to see if the user exists on the system.  If the user exists and the script is being run as root, it will drop privileges to that user.